Functions moved to lib.py, now imported as library functions
