+++ thoth/glyph/__init__.py
@@ -17,6 +17,11 @@
 
 """Generate CHANGELOG entries out of commit messages using AI/ML techniques."""
 
+from .lib import classify_message
+from .lib import classify_messages
+from .lib import classify_by_date
+from .lib import classify_by_tag
+
 __author__ = "Tushar Sharma <tussharm@redhat.com>"
 __title__ = "glyph"
 __version__ = "0.0.0"
